Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240619/합과 평균/sum-and-avg.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-    double a, b;
+    int a, b;
     cin >> a >> b;
-    cout << a+b << " " << (a+b)/2;
+    cout << a+b << " " << (double)(a+b)/2;
     return 0;
 }
